// src/app/pages/cron-quartz/models/quartzModels.ts
export class QuartzModel {
  jobTitle?: any;
  cronExpression?: any;
  createdBy?: any;
  expression?: any;
  id?: any;
  createdDate?: any;
  serviceUrlId?: any;
  flavour?: any;
  flavourType?: any;
  jobDescription?: any;
  cronExpressionPayload?: any;
  jobStartDate?: string;
  jobEndDate?: string;
  scheduledJobStatus?: any;
  recordStatus?: any;
  updatedBy?: any;
  cronTitle?: any;
  updatedDate?: any;
  startDate?: string;
  endDate?: string;
  cronExpPayload?: string;
  triggeredUrlId?: any;
  cronDescription?: string;
  categoryId: string;
}